Refactor: 프로젝션을 위한 ReservationDTO 위치 변경 (#53)
