+++ src/main/java/com/livable/server/reservation/repository/ReservationQueryRepository.java
@@ -1,13 +1,13 @@
 package com.livable.server.reservation.repository;
 
 import com.livable.server.entity.Reservation;
-import com.livable.server.invitation.dto.InvitationProjection;
+import com.livable.server.invitation.dto.InvitationResponse;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 public interface ReservationQueryRepository {
-    List<InvitationProjection.ReservationDTO> findReservationsByCompanyId(Long companyId);
+    List<InvitationResponse.ReservationDTO> findReservationsByCompanyId(Long companyId);
     List<Reservation> findReservationsByCommonPlaceIdAndStartDateAndEndDate(
             Long commonPlaceId, LocalDateTime startDateTime, LocalDateTime endDateTime);
 }
